+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Added
  - Armory logger
 
+### Fixed
+ - Lack of banking channel 2
+
 ## [1.3]
 ### Added
  - Access log to log unauthorized attempts to run commands